# server/docs/swagger.yaml
basePath: /
definitions:
...
  api.gstreamerProbeInfoDoc:
    properties:
      Container:
        type: string
      DurationNS:
        type: integer
      FileSize:
        type: integer
      Tracks:
        items:
          $ref: '#/definitions/api.gstreamerTrackDoc'
        type: array
    type: object
...
  api.gstreamerTrackDoc:
    properties:
      CapsName:
        type: string
      Channels:
        type: integer
      Codec:
        type: string
      FrameRateDen:
        type: integer
      FrameRateNum:
        type: integer
      Height:
        type: integer
      Index:
        type: integer
      Language:
        type: string
      PadName:
        type: string
      Rate:
        type: integer
      Title:
        type: string
      Type:
        type: string
      Width:
        type: integer
    type: object
...
  models.TorrentDetails:
    properties:
      audioQuality:
        type: integer
      categories:
        type: string
      createDate:
        type: string
      hash:
        type: string
      imdbid:
        type: string
      link:
        type: string
      magnet:
        type: string
      name:
        type: string
      names:
        items:
          type: string
        type: array
      peer:
        type: integer
      seed:
        type: integer
      size:
        type: string
      title:
        type: string
      tracker:
        type: string
      videoQuality:
        type: integer
      year:
        type: integer
    type: object
...
info:
  contact: {}
  description: Torrent streaming server.
  license:
    name: GPL 3.0
  title: Swagger Torrserver API
  version: '{version.Version}'
paths:
...
  /ffp/{hash}/{id}:
    get:
      description: Gather informations using ffprobe.
      parameters:
      - description: Torrent hash
        in: path
        name: hash
        required: true
        type: string
      - description: File index in torrent
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Data returned from ffprobe
      summary: Gather informations using ffprobe
      tags:
      - API
...
  /gst/{hash}/probe:
    get:
      description: Requires a `-gst` build. Uses gst-discoverer to inspect container
        and codec information.
      parameters:
      - description: Torrent infohash
        in: path
        name: hash
        required: true
        type: string
      - description: File index in torrent
        in: query
        name: index
        type: string
      - description: File index (alias)
        in: query
        name: id
        type: string
      - description: File index (alias)
        in: query
        name: fileID
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Media probe result
          schema:
            $ref: '#/definitions/api.gstreamerProbeInfoDoc'
        "400":
          description: Invalid source
          schema:
            additionalProperties:
              type: string
            type: object
        "502":
          description: Probe failed
          schema:
            type: string
        "504":
          description: Probe timed out
          schema:
            type: string
      summary: Probe torrent media tracks
      tags:
      - GStreamer
...
  /jacred/search:
    get:
      description: Searches the configured JacRed aggregator.
      parameters:
      - description: JacRed query
        in: query
        name: query
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: JacRed torrent search result(s)
          schema:
            items:
              $ref: '#/definitions/models.TorrentDetails'
            type: array
      summary: Makes a JacRed search
      tags:
      - API
...
  /play/{hash}/{id}:
    get:
      description: Play given torrent referenced by infohash and file id.
      parameters:
      - description: Torrent infohash
        in: path
        name: hash
        required: true
        type: string
      - description: File index in torrent
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: Torrent data
      summary: Play given torrent by infohash
      tags:
      - API
  /playlist:
    get:
      description: Get HTTP link of torrent in M3U list.
      parameters:
      - description: Torrent hash
        in: query
        name: hash
        required: true
        type: string
      - description: From last play file
        in: query
        name: fromlast
        type: boolean
      - description: Start the playlist at this file id (overrides fromlast)
        in: query
        name: index
        type: integer
      produces:
      - audio/x-mpegurl
      responses:
        "200":
          description: OK
          schema:
            type: file
      summary: Get HTTP link of torrent in M3U list
      tags:
      - API
...
  /stream:
    get:
      description: Multi usage endpoint.
      parameters:
      - description: Magnet/hash/link to torrent
        in: query
        name: link
        required: true
        type: string
      - description: File index in torrent
        in: query
        name: index
        type: string
      - description: Should preload torrent
        in: query
        name: preload
        type: string
      - description: Get statistics from torrent
        in: query
        name: stat
        type: string
      - description: Should save torrent
        in: query
        name: save
        type: string
      - description: Get torrent as M3U playlist
        in: query
        name: m3u
        type: string
      - description: Get M3U from last played file
        in: query
        name: fromlast
        type: string
      - description: Start stream torrent
        in: query
        name: play
        type: string
      - description: Set title of torrent
        in: query
        name: title
        type: string
      - description: Set poster link of torrent
        in: query
        name: poster
        type: string
      - description: 'Set category of torrent, used in web: movie, tv, music, other'
        in: query
        name: category
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: Data returned according to query
      summary: Multi usage endpoint
      tags:
      - API
...
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"